Fix stacktrace when using --extract if the cert has no CN.
Ubuntu's ships self-signed "snakeoil" as part of the OVMF package.
They're useful for simple testing. The certs as found in
current version (2022.02-3ubuntu0.22.04.1) do not have a CN set.

This is probably an edge case for "real" certs, but if here is
no CN found use the sha256sum of the DER encoded certificate in
place of the CN.