kernel-bootcfg based kernel-install script for direct UKI load from shim
'kernel-bootcfg' utility allows to add and remove UKIs booted directly from
shim from UEFI boot variables. Add a kernel-install script to automate the
process. Also, support A/B UKI loading with a newly introduced
kernel-bootcfg-boot-successful.service.

Note, minor complication comes from the fact that the standard systemd's
'90-uki-copy.install' puts UKI to $BOOT/EFI/Linux/ and this may or may not
correspond to ESP/EFI/Linux. E.g. In case ESP is not used to host /boot, it
may get mounted to /boot/efi/ and '90-uki-copy.install' will put the kernel
to /boot/EFI/Linux and not to /boot/efi/EFI/Linux. shim can't boot UKI in
case it is not on ESP. The behavior of '90-uki-copy.install' is intentional
and complies to the Boot Loader Specification:

"""
Operating systems should place unified EFI kernels only in the $BOOT
partition. Boot loaders should also look in the /EFI/Linux/ of the ESP — if
it is different from $BOOT — and present a merged list of menu entries from
both partitions.
"""

Support this in 99-uki-uefi-setup.install by doing an additional copy of
the UKI onto the ESP if needed.

Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>

[ kraxel: throw errors in case boot counting is active ]
[ kraxel: minor fixes ]

Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>